backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
distro==1.9.0
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
joblib==1.5.3
numpy==2.4.4
openai==2.33.0
pandas==3.0.2
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
uvicorn==0.46.0
